+++ HMS.Application/Services/RoomService.cs
@@ -22,14 +22,17 @@ namespace HMS.Application.Services
 
             return MapToPagedResponseDto(rooms, parameters);
         }
+
+        
+
         public async Task<RoomForGettingDto> GetRoomByIdAsync(int roomId)
         {
-            if(roomId == null)
+            if (roomId == null)
                 throw new BadRequestException("Room Id is required");
 
             var room = roomRepository.GetAsync(r => r.RoomId == roomId);
 
-            if(room is null)
+            if (room is null)
                 throw new NotFoundException($"Room with Id {roomId} not found");
 
             return mapper.Map<RoomForGettingDto>(room);
@@ -37,8 +40,14 @@ namespace HMS.Application.Services
         }
         public async Task<int> CreateRoomAsync(RoomForCreatingDto model)
         {
-            if(model is null)
+            if (model is null)
                 throw new BadRequestException("Room model is required");
+
+            //if room is preserved same room cant be preserved again
+            var existingRoom = await roomRepository.GetAsync(r => r.Name == model.Name);
+
+            if (existingRoom is not null)
+                throw new BadRequestException($"Room with name {model.Name} already exists");
 
             var room = mapper.Map<Room>(model);
             await roomRepository.AddAsync(room);
@@ -48,12 +57,12 @@ namespace HMS.Application.Services
         }
         public async Task<RoomForGettingDto> UpdateRoomAsync(RoomForUpdatingDto model)
         {
-            if(model is null)
+            if (model is null)
                 throw new BadRequestException("Room model is required");
 
             var room = await roomRepository.GetAsync(r => r.RoomId == model.RoomId);
 
-            if(room is null)
+            if (room is null)
                 throw new NotFoundException($"Room with Id {model.RoomId} not found");
 
             mapper.Map(model, room);
@@ -66,13 +75,21 @@ namespace HMS.Application.Services
 
         public async Task DeleteRoomAsync(int roomId)
         {
-            if(roomId == null)
+            if (roomId == null)
                 throw new BadRequestException("Room Id is required");
 
             var room = await roomRepository.GetAsync(r => r.RoomId == roomId);
 
-            if(room is null)
+            if (room is null)
                 throw new NotFoundException($"Room with Id {roomId} not found");
+
+            var today = DateTime.UtcNow.Date;
+
+            bool hasActiveOrFutureReservations = room.ReservationRooms
+            .Any(rr => rr.Reservation.CheckOutDate >= today);
+
+            if(hasActiveOrFutureReservations)
+                throw new BadRequestException($"Room with Id {roomId} has active or future reservations and cannot be deleted");
 
             roomRepository.Remove(room);
             await roomRepository.SaveAsync();
@@ -105,5 +122,5 @@ namespace HMS.Application.Services
 
 
         }
-    } 
+    }
 }
